added more service and dao objects
git-svn-id: https://svn.kuali.org/repos/kpme/trunk@290 6c5f4195-72ea-6250-e7e4-e498f552a5df

+++ src/main/java/org/kuali/hr/job/Job.java
@@ -1,5 +1,6 @@
 package org.kuali.hr.job;
 
+import java.math.BigInteger;
 import java.sql.Date;
 import java.sql.Timestamp;
 import java.util.LinkedHashMap;
@@ -9,7 +10,7 @@ import org.kuali.rice.kns.bo.PersistableBusinessObjectBase;
 public class Job  extends PersistableBusinessObjectBase {
 
     /**
-     * 
+     *
      */
     private static final long serialVersionUID = 1L;
     private String location;
@@ -19,21 +20,21 @@ public class Job  extends PersistableBusinessObjectBase {
     private int standardHours;
     private Long jobId;
     private String principalId;
-    private Long jobNumber;
+    private BigInteger jobNumber;
     private Date effectiveDate;
     private String deptId;
     private Long payCalendarId;
     private String tkSalGroup;
     private Timestamp timestamp;
     private Boolean active;
-    
+
     @SuppressWarnings("unchecked")
     @Override
     protected LinkedHashMap toStringMapper() {
 	LinkedHashMap<String, Object> toStringMap = new LinkedHashMap<String,Object>();
-	
+
 	toStringMap.put("principalId", principalId);
-	
+
 	return toStringMap;
     }
 
@@ -95,12 +96,12 @@ public class Job  extends PersistableBusinessObjectBase {
     }
 
 
-    public Long getJobNumber() {
+    public BigInteger getJobNumber() {
         return jobNumber;
     }
 
 
-    public void setJobNumber(Long jobNumber) {
+    public void setJobNumber(BigInteger jobNumber) {
         this.jobNumber = jobNumber;
     }
 
@@ -113,7 +114,7 @@ public class Job  extends PersistableBusinessObjectBase {
     public void setEffectiveDate(Date effectiveDate) {
         this.effectiveDate = effectiveDate;
     }
-    
+
 
     public Long getPayCalendarId() {
         return payCalendarId;
